refactor: 도메인과 엔티티 패키지 분리
