+++ backend/src/main/java/reviewme/highlight/repository/HighlightRepository.java
@@ -4,7 +4,7 @@ import java.util.Collection;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import reviewme.highlight.domain.Highlight;
+import reviewme.highlight.entity.Highlight;
 
 public interface HighlightRepository extends JpaRepository<Highlight, Long> {
