+++ backend/src/main/java/reviewme/highlight/service/mapper/HighlightMapper.java
@@ -7,10 +7,10 @@ import java.util.Map.Entry;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import reviewme.highlight.domain.Highlight;
+import reviewme.highlight.entity.Highlight;
 import reviewme.highlight.domain.HighlightLine;
 import reviewme.highlight.domain.HighlightLines;
-import reviewme.highlight.domain.HighlightRange;
+import reviewme.highlight.entity.HighlightRange;
 import reviewme.highlight.service.dto.HighlightIndexRangeRequest;
 import reviewme.highlight.service.dto.HighlightRequest;
 import reviewme.highlight.service.dto.HighlightedLineRequest;
