+++ backend/src/test/java/reviewme/highlight/service/HighlightServiceTest.java
@@ -10,6 +10,10 @@ import static reviewme.fixture.TemplateFixture.템플릿;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import reviewme.fixture.QuestionFixture;
+import reviewme.fixture.ReviewGroupFixture;
+import reviewme.highlight.entity.Highlight;
+import reviewme.highlight.entity.HighlightRange;
 import reviewme.highlight.domain.Highlight;
 import reviewme.highlight.domain.HighlightRange;
 import reviewme.highlight.repository.HighlightRepository;
